test selector and duplicate tests clean up
